requirements.txt
blinker==1.4
Brlapi==0.8.2
cassandra-driver==3.25.0
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==7.1.2
colorama==0.4.4
command-not-found==0.3
cryptography==3.3.2
cupshelpers==1.0
datashape==0.5.2
dbus-python==1.2.16
defer==1.0.6
distro==1.5.0
dnspython==2.1.0
geomet==0.2.1.post1
gpg===1.14.0-unknown
greenlet==1.1.2
hidpidaemon==18.4.6
httplib2==0.18.1
idna==2.10
jeepney==0.6.0
kernelstub==3.1.4
keyring==22.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
louis==3.16.0
macaroonbakery==1.3.1
multipledispatch==0.6.0
mysql-connector-python==8.0.27
netifaces==0.10.9
networkx==2.6.3
numpy==1.21.4
oauthlib==3.1.0
odo==0.5.0
olefile==0.46
packaging==21.2
pandas==1.3.4
pdfkit==1.0.0
pexpect==4.8.0
Pillow==8.1.2
pip-review==1.1.0
pop-transition==1.1.2
protobuf==3.12.4
psycopg==3.0.5
psycopg2-binary==2.9.2
pycairo==1.16.2
pycups==2.0.1
pydbus==0.6.0
PyGObject==3.38.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==4.0
PyNaCl==1.4.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-sip==12.8.1
pyRFC3339==1.1
python-apt===2.2.0-ubuntu0.21.04.1
python-dateutil==2.8.2
python-debian==0.1.39
python-xlib==0.29
pytz==2021.1
pyudev==0.22.0
pyxdg==0.27
PyYAML==5.3.1
repolib==1.5.2
repoman==1.4.0
reportlab==3.5.66
requests==2.25.1
scipy==1.7.3
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sessioninstaller==0.0.0
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.4.27
systemd-python==234
toolz==0.11.2
tqdm==4.62.3
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.26.2
vboxapi==1.0
wadllib==1.3.5
xdg==5
xkit==0.0.0
